code file refactoring, move data formatter as standalone file
